requirements.txt
aiohttp
aiosqlite
aiounittest
pytest-cov
pillow